Added feature for disabling a workflow.
This implements a lifecycle for the workflows themselves.

+++ classes/manager/workflow_manager.php
@@ -21,6 +21,7 @@
  * @copyright  2017 Tobias Reischmann WWU
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+
 namespace tool_lifecycle\manager;
 
 use tool_lifecycle\entity\trigger_subplugin;
@@ -32,7 +33,8 @@ defined('MOODLE_INTERNAL') || die();
 class workflow_manager {
 
     /**
-     * Remove a workflow from the database.
+     * Persists a workflow to the database.
+     *
      * @param workflow $workflow
      */
     public static function insert_or_update(workflow &$workflow) {
@@ -47,18 +49,51 @@ class workflow_manager {
     }
 
     /**
-     * Persists a workflow to the database.
+     * Remove a workflow from the database.
+     *
+     * @param int $workflowid id of the workflow
+     * @param boolean $hard if set, will remove the workflow without checking if it's removable! Mainly for testing.
+     */
+    public static function remove($workflowid, $hard = false) {
+        global $DB;
+        if ($hard || self::is_removable($workflowid)) {
+            trigger_manager::remove_instances_of_workflow($workflowid);
+            step_manager::remove_instances_of_workflow($workflowid);
+            $DB->delete_records('tool_lifecycle_workflow', array('id' => $workflowid));
+        }
+    }
+
+    /**
+     * Disables a workflow
+     *
      * @param int $workflowid id of the workflow
      */
-    public static function remove($workflowid) {
-        global $DB;
-        trigger_manager::remove_instances_of_workflow($workflowid);
-        step_manager::remove_instances_of_workflow($workflowid);
-        $DB->delete_records('tool_lifecycle_workflow', array('id' => $workflowid));
+    public static function disable($workflowid) {
+        $workflow = self::get_workflow($workflowid);
+        if ($workflow && self::is_disableable($workflowid)) { // @TODO Notify user if not.
+            $workflow->active = false;
+            // $workflow->timeactive = null; @TODO Necessary? Why do we have active and timeactive?
+            $workflow->sortindex = null;
+            $workflow->timedeactive = time();
+            self::insert_or_update($workflow);
+        }
+    }
+
+    /**
+     * Deletes all running processes of given workflow
+     *
+     * @param int $workflowid id of the workflow
+     */
+    public static function abortprocesses($workflowid) {
+        $processes = process_manager::get_processes_by_workflow($workflowid);
+        foreach ($processes as $process) {
+            process_manager::rollback_process($process);
+        }
     }
 
     /**
      * Returns a workflow instance if one with the is is available.
+     *
      * @param int $workflowid id of the workflow
      * @return workflow|null
      */
@@ -75,6 +110,7 @@ class workflow_manager {
 
     /**
      * Returns all active workflows.
+     *
      * @return workflow[]
      */
     public static function get_active_workflows() {
@@ -90,6 +126,7 @@ class workflow_manager {
 
     /**
      * Returns all active automatic workflows.
+     *
      * @return workflow[]
      */
     public static function get_active_automatic_workflows() {
@@ -105,12 +142,13 @@ class workflow_manager {
 
     /**
      * Returns triggers of active manual workflows.
+     *
      * @return trigger_subplugin[]
      */
     public static function get_active_manual_workflow_triggers() {
         global $DB;
-        $sql = 'SELECT t.* FROM {tool_lifecycle_workflow} w JOIN {tool_lifecycle_trigger} t ON t.workflowid = w.id'.
-        ' WHERE w.active = ? AND w.manual = ?';
+        $sql = 'SELECT t.* FROM {tool_lifecycle_workflow} w JOIN {tool_lifecycle_trigger} t ON t.workflowid = w.id' .
+            ' WHERE w.active = ? AND w.manual = ?';
         $records = $DB->get_records_sql($sql, array(true, true));
         $result = array();
         foreach ($records as $record) {
@@ -122,6 +160,7 @@ class workflow_manager {
     /**
      * Returns tools for all active manual workflows.
      * You need to check the capability based on course and user before diplaying it.
+     *
      * @return manual_trigger_tool[] list of tools, available in the whole system.
      */
     public static function get_manual_trigger_tools_for_active_workflows() {
@@ -136,6 +175,7 @@ class workflow_manager {
 
     /**
      * Activate a workflow
+     *
      * @param int $workflowid id of the workflow
      */
     public static function activate_workflow($workflowid) {
@@ -165,11 +205,13 @@ class workflow_manager {
 
     /**
      * Handles an action of the subplugin_settings.
+     *
      * @param string $action action to be executed
      * @param int $workflowid id of the workflow
      */
     public static function handle_action($action, $workflowid) {
         global $OUTPUT;
+        $confirm = optional_param('confirm', 0, PARAM_BOOL);
         if ($action === ACTION_WORKFLOW_ACTIVATE) {
             self::activate_workflow($workflowid);
         }
@@ -182,19 +224,48 @@ class workflow_manager {
         if ($action === ACTION_WORKFLOW_DUPLICATE) {
             self::duplicate_workflow($workflowid);
         }
+        if ($action === ACTION_WORKFLOW_DISABLE) {
+            if (confirm_sesskey()) {
+                self::disable($workflowid);
+            }
+        }
+        if ($action === ACTION_WORKFLOW_ABORTDISABLE) {
+            if (confirm_sesskey()) {
+                self::disable($workflowid);
+                self::abortprocesses($workflowid);
+            }
+        }
+        if ($action === ACTION_WORKFLOW_ABORT) {
+            if (confirm_sesskey()) {
+                self::abortprocesses($workflowid);
+            }
+        }
         if ($action === ACTION_WORKFLOW_DELETE) {
-            if (self::is_active($workflowid)) {
-                echo $OUTPUT->notification(get_string('active_workflow_not_removeable', 'tool_lifecycle')
-                    , 'warning');
-
-            } else {
+            // Check workflow wasn't already deleted, in case someone refreshes the page.
+            if (self::get_workflow($workflowid) &&
+                self::is_removable($workflowid) &&
+                confirm_sesskey()) {
                 self::remove($workflowid);
+            } else {
+                echo $OUTPUT->notification(get_string('workflow_not_removeable', 'tool_lifecycle')
+                    , 'warning'); // @todo these notifications aren't shown properly currently
             }
         }
     }
 
+    private static function render_demand_confirm($action, $workflowid, $message) {
+        global $OUTPUT, $PAGE;
+        $yesurl = new \moodle_url($PAGE->url, array('workflowid' => $workflowid, 'action' => $action, 'sesskey' => sesskey(), 'confirm' => 1));
+        $nourl = new \moodle_url('/admin/tool/lifecycle/adminsettings.php');
+        $output = $OUTPUT->header();
+        $output .= $OUTPUT->confirm($message, $yesurl, $nourl);
+        $output .= $OUTPUT->footer();
+        echo $output;
+    }
+
     /**
      * Changes the sortindex of a workflow by swapping it with another.
+     *
      * @param int $workflowid id of the workflow
      * @param bool $up tells if the workflow should be set up or down
      */
@@ -238,6 +309,7 @@ class workflow_manager {
     /**
      * Checks if the workflow definition is valid.
      * The main purpose of this function is, to check if a trigger definition exists and if this definition is complete.
+     *
      * @param $workflowid int id of the workflow.
      * @return bool true, if the definition is valid.
      */
@@ -251,6 +323,7 @@ class workflow_manager {
 
     /**
      * Checks if the workflow is active.
+     *
      * @param $workflowid int id of the workflow.
      * @return bool true, if the workflow is active.
      */
@@ -260,7 +333,22 @@ class workflow_manager {
     }
 
     /**
+     * Checks if the workflow is deactive.
+     *
+     * @param $workflowid int id of the workflow.
+     * @return bool true, if the workflow was deactivated.
+     */
+    public static function is_deactivated($workflowid) {
+        $workflow = self::get_workflow($workflowid);
+        if ($workflow->timedeactive) {
+            return true;
+        }
+        return false;
+    }
+
+    /**
      * Creates a workflow with a specific title. Is used to create preset workflows for trigger plugins.
+     *
      * @param $title string title of the workflow. Usually the pluginname of the trigger.
      * @return workflow the created workflow.
      */
@@ -274,6 +362,7 @@ class workflow_manager {
 
     /**
      * Duplicates a workflow including its trigger, all its steps and their settings.
+     *
      * @param $workflowid int id of the workflow to copy.
      * @return workflow the created workflow.
      */
@@ -292,4 +381,62 @@ class workflow_manager {
         return $newworkflow;
     }
 
+    /**
+     * Checks if it should be possible to disable a workflow
+     *
+     * @param $workflowid
+     * @return bool
+     */
+    public static function is_disableable($workflowid) {
+        $trigger = trigger_manager::get_triggers_for_workflow($workflowid);
+        if (!empty($trigger)) {
+            $lib = lib_manager::get_trigger_lib($trigger[0]->subpluginname);
+        }
+        if (!isset($lib) || $lib->has_multiple_instances()) {
+            return true;
+        }
+        return false;
+    }
+
+    /**
+     * Workflows should only be editable if never been activated before
+     *
+     * @param $workflowid
+     * @return bool
+     */
+    public static function is_editable($workflowid) {
+        if (self::is_active($workflowid) ||
+            self::is_deactivated($workflowid)) {
+            return false;
+        }
+        return true;
+    }
+
+    /**
+     * Workflows should only be abortable if disabled but some processes are still running
+     *
+     * @param $workflowid
+     * @return bool
+     */
+    public static function is_abortable($workflowid) {
+        $countprocesses = process_manager::count_processes_by_workflow($workflowid);
+        if ($countprocesses > 0) {
+            return true;
+        }
+        return false;
+    }
+
+    /**
+     * Workflows should only be removable if disableable and no more processes are running
+     *
+     * @param $workflowid
+     * @return bool
+     */
+    public static function is_removable($workflowid) {
+        $countprocesses = process_manager::count_processes_by_workflow($workflowid);
+        if (self::is_disableable($workflowid) && $countprocesses == 0) {
+            return true;
+        }
+        return false;
+    }
 }
